[MOD] Refactor MemberInfoForm to be containerless
