+++ src/app/components/addMember/addMemberComponent.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { browserHistory, Link } from 'react-router';
 
-import MemberInfoForm from '../memberInfoForm/memberInfoFormContainer';
+import MemberInfoForm from '../../forms/memberInfoForm';
 
 export default ({ updateMember }) => (
   <div className="container add-member">
@@ -9,6 +9,6 @@ export default ({ updateMember }) => (
     <h1>Add a team member</h1>
     <h4 className="subtitle">Set email, location and role.</h4>
     <div className="divider"></div>
-    <MemberInfoForm initialValues={{role: 'regular'}} onSubmit={ member => { updateMember(member); browserHistory.push('/') } } />
+    <MemberInfoForm editting={false} initialValues={{role: 'regular'}} onSubmit={ member => { updateMember(member); browserHistory.push('/') } } />
   </div>
 )
